Organize the folders with subdirectories
Because the number of classes tend to grow fast in connector addons, organize
directories using a hierarchy:

- models
|-- model_name
|---- common.py: hold Models, Binders, Adapters (what is shared for importers and exporters)
|---- importer.py: hold functions and classes related to imports
|---- exporter.py: hold functions and classes related to exports
