+++ magentoerpconnect_catalog_simple/models/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from . import magento_model
+from . import magento_attribute_set
