+++ magentoerpconnect_catalog_simple/models/magento_attribute_set/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+
+from . import common
+from . import importer
