+++ magentoerpconnect_catalog_simple/models/magento_model/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from . import common
