+++ magentoerpconnect_catalog_simple/views/magento_model_view.xml
